requirements.txt
Django==5.0.6
djangorestframework==3.15.0
flake8
psycopg2-binary==2.9.6